// Features/reports/Types/ILatenessForReport.ts
export interface ILatenessForReport {
    identifier: string;
    idLateness: number;
    idEmployee: number;
    fullName: string;
    costCenter: string;
    salary: number;
    percentage: number;
    concept: string;
    hourAmount: number;
    amount: number;
    date: string;
    payrollName: string;
    idPayrollPay: number;
    isPaid: string;
    idPosition: number;
    position: string;
    idDepartment: number;
    department: string;
    accountNumber: string;
    idCostCenter: number;
    idCompany: number;
}